Replaced Character.java with RefactoredCharacter.java
